# src/main/resources/openapi.yaml
openapi: 3.1.1
info:
  title: Distributed Messenger API
  version: 1.0.0
servers:
- description: API Version 2
  url: /api/v2
security:
- bearerAuth: []
...
paths:
...
  /users:
    get:
      operationId: getAllUsers
      parameters:
      - explode: true
        in: query
        name: username
        required: false
        schema:
          type: string
        style: form
      - description: Номер страницы
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - description: Количество элементов на странице
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 25
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
          description: Страница со списком пользователей.
        "403":
          description: Доступ запрещен.
      summary: Получить список всех пользователей (Admin)
      tags:
      - Users
  /users/{userId}:
    delete:
      operationId: deleteUser
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Пользователь удален.
        "403":
          description: Доступ запрещен.
        "404":
          description: Пользователь не найден.
      summary: Удалить пользователя по ID (Admin)
      tags:
      - Users
    get:
      operationId: getUserById
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Профиль пользователя.
        "404":
          description: Пользователь не найден.
      summary: Получить пользователя по ID
      tags:
      - Users
    patch:
      operationId: updateUser
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Пользователь обновлен.
        "403":
          description: Доступ запрещен.
        "404":
          description: Пользователь не найден.
      summary: Обновить пользователя по ID
      tags:
      - Users
...
components:
  schemas:
    User:
      example:
        profileSettingsId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        role: UNAUTHORIZED_USER
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        appSettingsId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        username: username
      properties:
        id:
          format: uuid
          type: string
        username:
          type: string
        role:
          enum:
          - UNAUTHORIZED_USER
          - USER
          - MODERATOR
          - ADMINISTRATOR
          type: string
        profileSettingsId:
          format: uuid
          type: string
        appSettingsId:
          format: uuid
          type: string
      required:
      - id
      - role
      - username
...
    UserUpdate:
      example:
        role: USER
        username: username
      properties:
        username:
          type: string
        role:
          enum:
          - USER
          - MODERATOR
          - ADMINISTRATOR
          type: string
...
    PaginationInfo:
      description: Метаданные для постраничной навигации.
      example:
        totalItems: 0
        totalPages: 6
        limit: 5
        currentPage: 1
      properties:
        totalItems:
          type: integer
        totalPages:
          type: integer
        currentPage:
          type: integer
        limit:
          type: integer
    PaginatedUsers:
      example:
        pagination:
          totalItems: 0
          totalPages: 6
          limit: 5
          currentPage: 1
        items:
        - profileSettingsId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          role: UNAUTHORIZED_USER
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          appSettingsId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          username: username
        - profileSettingsId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          role: UNAUTHORIZED_USER
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          appSettingsId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          username: username
      properties:
        items:
          items:
            $ref: "#/components/schemas/User"
          type: array
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http